refactor : Engine modularization
