+++ engine/core/function/utils/cosineSimilarity.ts
@@ -1,0 +1,6 @@
+export function calculateCosineSimilarity(vectorA: number[], vectorB: number[]): number {
+    const dotProduct = vectorA.reduce((sum, value, index) => sum + value * vectorB[index], 0);
+    const magnitudeA = Math.sqrt(vectorA.reduce((sum, value) => sum + value * value, 0));
+    const magnitudeB = Math.sqrt(vectorB.reduce((sum, val) => sum + val * val, 0));
+    return dotProduct / (magnitudeA * magnitudeB);
+}
